requirements.txt
streamlit
pymongo
google-generativeai
bson
